refactored another big integration test into new server format
