+++ tests/test_main_integration.py
@@ -208,318 +208,10 @@ def test_full_datapath(
     assert is_queue_eventually_empty(da_error_queue) is True
 
 
-@pytest.mark.slow
-@pytest.mark.timeout(INTEGRATION_TEST_TIMEOUT)
-@freeze_time(
-    datetime.datetime(
-        year=2020, month=6, day=15, hour=14, minute=19, second=55, microsecond=313309
-    )
-)
-def test_system_states_and_recorded_metadata_with_update_to_file_writer_directory(
-    patched_xem_scripts_folder,
-    patched_firmware_folder,
-    test_process_manager,
-    test_client,
-    fully_running_app_from_main_entrypoint,
-    mocker,
-):
-    expected_time = datetime.datetime(
-        year=2020, month=6, day=15, hour=14, minute=19, second=55, microsecond=313309
-    )
-    mocker.patch.object(
-        main,
-        "_get_timestamp_of_acquisition_sample_index_zero",
-        return_value=expected_time,
-    )
-    expected_timestamp = "2020_06_15_141955"
-
-    _ = fully_running_app_from_main_entrypoint(["--skip-mantarray-boot-up"])
-    wait_for_subprocesses_to_start()
-    response = requests.get(f"{get_api_endpoint()}system_status")
-    assert response.status_code == 200
-    assert response.json()["ui_status_code"] == str(
-        SYSTEM_STATUS_UUIDS[SERVER_READY_STATE]
-    )
-
-    with tempfile.TemporaryDirectory() as expected_recordings_dir:
-        response = requests.get(
-            f"{get_api_endpoint()}update_settings?customer_account_uuid=curi&recording_directory={expected_recordings_dir}"
-        )
-        assert response.status_code == 200
-
-        response = requests.get(f"{get_api_endpoint()}boot_up")
-        assert response.status_code == 200
-        response = requests.get(f"{get_api_endpoint()}system_status")
-        assert response.status_code == 200
-        assert response.json()["ui_status_code"] == str(
-            SYSTEM_STATUS_UUIDS[INSTRUMENT_INITIALIZING_STATE]
-        )
-        assert system_state_eventually_equals(CALIBRATION_NEEDED_STATE, 3) is True
-
-        response = requests.get(f"{get_api_endpoint()}start_calibration")
-        assert response.status_code == 200
-        response = requests.get(f"{get_api_endpoint()}system_status")
-        assert response.status_code == 200
-        assert response.json()["ui_status_code"] == str(
-            SYSTEM_STATUS_UUIDS[CALIBRATING_STATE]
-        )
-        assert (
-            system_state_eventually_equals(CALIBRATED_STATE, CALIBRATED_WAIT_TIME)
-            is True
-        )
-
-        response = requests.get(f"{get_api_endpoint()}start_managed_acquisition")
-        assert response.status_code == 200
-        response = requests.get(f"{get_api_endpoint()}system_status")
-        assert response.status_code == 200
-        assert response.json()["ui_status_code"] == str(
-            SYSTEM_STATUS_UUIDS[BUFFERING_STATE]
-        )
-        assert (
-            system_state_eventually_equals(
-                LIVE_VIEW_ACTIVE_STATE, LIVE_VIEW_ACTIVE_WAIT_TIME
-            )
-            is True
-        )
-
-        expected_barcode = GENERIC_START_RECORDING_COMMAND[
-            "metadata_to_copy_onto_main_file_attributes"
-        ][PLATE_BARCODE_UUID]
-        start_recording_time_index = 960
-        response = requests.get(
-            f"{get_api_endpoint()}start_recording?barcode={expected_barcode}&time_index={start_recording_time_index}&is_hardware_test_recording=False"
-        )
-        assert response.status_code == 200
-        response = requests.get(f"{get_api_endpoint()}system_status")
-        assert response.status_code == 200
-        assert response.json()["ui_status_code"] == str(
-            SYSTEM_STATUS_UUIDS[RECORDING_STATE]
-        )
-
-        time.sleep(3)  # Tanner (6/15/20): This allows data to be written to files
-
-        response = requests.get(f"{get_api_endpoint()}stop_recording")
-        assert response.status_code == 200
-        response = requests.get(f"{get_api_endpoint()}system_status")
-        assert response.status_code == 200
-        assert response.json()["ui_status_code"] == str(
-            SYSTEM_STATUS_UUIDS[LIVE_VIEW_ACTIVE_STATE]
-        )
-
-        response = requests.get(f"{get_api_endpoint()}stop_managed_acquisition")
-        assert response.status_code == 200
-        response = requests.get(f"{get_api_endpoint()}system_status")
-        assert response.status_code == 200
-        assert (
-            system_state_eventually_equals(
-                CALIBRATED_STATE, STOP_MANAGED_ACQUISITION_WAIT_TIME
-            )
-            is True
-        )
-
-        test_process_manager.soft_stop_processes()
-
-        fw_process = test_process_manager.get_file_writer_process()
-        fw_process.close_all_files()
-        actual_set_of_files = set(
-            os.listdir(
-                os.path.join(
-                    expected_recordings_dir, f"{expected_barcode}__{expected_timestamp}"
-                )
-            )
-        )
-        assert len(actual_set_of_files) == 24
-
-        # Tanner (6/15/20): Processes must be joined to avoid h5 errors with reading files, so clearing queues manually here in order to join
-        response = requests.get(f"{get_api_endpoint()}get_available_data")
-        assert response.status_code == 200
-        response = requests.get(f"{get_api_endpoint()}get_available_data")
-        assert response.status_code == 200
-        test_process_manager.hard_stop_and_join_processes()
-
-        for row_idx in range(4):
-            for col_idx in range(6):
-                well_idx = col_idx * 4 + row_idx
-                with h5py.File(
-                    os.path.join(
-                        expected_recordings_dir,
-                        f"{expected_barcode}__{expected_timestamp}",
-                        f"{expected_barcode}__{expected_timestamp}__{chr(row_idx+65)}{col_idx+1}.h5",
-                    ),
-                    "r",
-                ) as this_file:
-                    this_file_attrs = this_file.attrs
-                    assert (
-                        bool(this_file_attrs[str(HARDWARE_TEST_RECORDING_UUID)])
-                        is False
-                    )
-                    assert (
-                        this_file_attrs[str(SOFTWARE_BUILD_NUMBER_UUID)]
-                        == COMPILED_EXE_BUILD_TIMESTAMP
-                    )
-                    assert (
-                        this_file_attrs["File Format Version"]
-                        == CURRENT_HDF5_FILE_FORMAT_VERSION
-                    )
-                    assert this_file_attrs[
-                        str(UTC_BEGINNING_DATA_ACQUISTION_UUID)
-                    ] == expected_time.strftime("%Y-%m-%d %H:%M:%S.%f")
-                    assert (
-                        this_file_attrs[str(START_RECORDING_TIME_INDEX_UUID)]
-                        == start_recording_time_index
-                    )
-                    assert this_file.attrs[
-                        str(UTC_BEGINNING_RECORDING_UUID)
-                    ] == expected_time.strftime("%Y-%m-%d %H:%M:%S.%f")
-                    assert this_file_attrs[str(UTC_FIRST_TISSUE_DATA_POINT_UUID)] == (
-                        expected_time
-                        + datetime.timedelta(
-                            seconds=(
-                                start_recording_time_index
-                                + WELL_24_INDEX_TO_ADC_AND_CH_INDEX[well_idx][1]
-                                * DATA_FRAME_PERIOD
-                            )
-                            / CENTIMILLISECONDS_PER_SECOND
-                        )
-                    ).strftime("%Y-%m-%d %H:%M:%S.%f")
-                    assert this_file_attrs[str(UTC_FIRST_REF_DATA_POINT_UUID)] == (
-                        expected_time
-                        + datetime.timedelta(
-                            seconds=(start_recording_time_index + DATA_FRAME_PERIOD)
-                            / CENTIMILLISECONDS_PER_SECOND
-                        )
-                    ).strftime("%Y-%m-%d %H:%M:%S.%f")
-                    assert this_file_attrs[str(USER_ACCOUNT_ID_UUID)] == str(
-                        CURI_BIO_USER_ACCOUNT_ID
-                    )
-                    assert this_file_attrs[str(CUSTOMER_ACCOUNT_ID_UUID)] == str(
-                        CURI_BIO_ACCOUNT_UUID
-                    )
-                    assert this_file_attrs[str(ADC_GAIN_SETTING_UUID)] == 16
-                    assert (
-                        this_file_attrs[str(ADC_TISSUE_OFFSET_UUID)]
-                        == FIFO_SIMULATOR_DEFAULT_WIRE_OUT_VALUE
-                    )
-                    assert (
-                        this_file_attrs[str(ADC_REF_OFFSET_UUID)]
-                        == FIFO_SIMULATOR_DEFAULT_WIRE_OUT_VALUE
-                    )
-                    assert (
-                        this_file_attrs[str(REFERENCE_VOLTAGE_UUID)]
-                        == REFERENCE_VOLTAGE
-                    )
-                    assert this_file_attrs[str(SLEEP_FIRMWARE_VERSION_UUID)] == "0.0.0"
-                    assert (
-                        this_file_attrs[str(MAIN_FIRMWARE_VERSION_UUID)]
-                        == RunningFIFOSimulator.default_firmware_version
-                    )
-                    assert (
-                        this_file_attrs[str(MANTARRAY_SERIAL_NUMBER_UUID)]
-                        == RunningFIFOSimulator.default_mantarray_serial_number
-                    )
-                    assert (
-                        this_file_attrs[str(MANTARRAY_NICKNAME_UUID)]
-                        == RunningFIFOSimulator.default_mantarray_nickname
-                    )
-                    assert (
-                        this_file_attrs[str(XEM_SERIAL_NUMBER_UUID)]
-                        == RunningFIFOSimulator.default_xem_serial_number
-                    )
-                    assert (
-                        this_file_attrs[str(SOFTWARE_RELEASE_VERSION_UUID)]
-                        == CURRENT_SOFTWARE_VERSION
-                    )
-
-                    assert (
-                        this_file_attrs[str(WELL_NAME_UUID)]
-                        == f"{chr(row_idx+65)}{col_idx+1}"
-                    )
-                    assert this_file_attrs["Metadata UUID Descriptions"] == json.dumps(
-                        str(METADATA_UUID_DESCRIPTIONS)
-                    )
-                    assert this_file_attrs[str(TOTAL_WELL_COUNT_UUID)] == 24
-                    assert this_file_attrs[str(WELL_ROW_UUID)] == row_idx
-                    assert this_file_attrs[str(WELL_COLUMN_UUID)] == col_idx
-                    assert (
-                        this_file_attrs[str(WELL_INDEX_UUID)] == row_idx + col_idx * 4
-                    )
-                    assert (
-                        this_file_attrs[str(TISSUE_SAMPLING_PERIOD_UUID)]
-                        == CONSTRUCT_SENSOR_SAMPLING_PERIOD
-                        * MICROSECONDS_PER_CENTIMILLISECOND
-                    )
-                    assert (
-                        this_file_attrs[str(REF_SAMPLING_PERIOD_UUID)]
-                        == REFERENCE_SENSOR_SAMPLING_PERIOD
-                        * MICROSECONDS_PER_CENTIMILLISECOND
-                    )
 
 
 
 
-@pytest.mark.slow
-@pytest.mark.timeout(INTEGRATION_TEST_TIMEOUT)
-def test_managed_acquisition_can_be_stopped_and_restarted_with_simulator(
-    patched_xem_scripts_folder,
-    patched_firmware_folder,
-    test_process_manager,
-    test_client,
-    fully_running_app_from_main_entrypoint,
-):
-    _ = fully_running_app_from_main_entrypoint()
-    wait_for_subprocesses_to_start()
-    assert system_state_eventually_equals(CALIBRATION_NEEDED_STATE, 5) is True
-
-    response = requests.get(f"{get_api_endpoint()}start_calibration")
-    assert response.status_code == 200
-    assert (
-        system_state_eventually_equals(CALIBRATED_STATE, CALIBRATED_WAIT_TIME) is True
-    )
-
-    # First run
-    response = requests.get(f"{get_api_endpoint()}start_managed_acquisition")
-    assert response.status_code == 200
-    assert (
-        system_state_eventually_equals(
-            LIVE_VIEW_ACTIVE_STATE, LIVE_VIEW_ACTIVE_WAIT_TIME
-        )
-        is True
-    )
-    response = requests.get(f"{get_api_endpoint()}stop_managed_acquisition")
-    assert response.status_code == 200
-    assert (
-        system_state_eventually_equals(
-            CALIBRATED_STATE, STOP_MANAGED_ACQUISITION_WAIT_TIME
-        )
-        is True
-    )
-
-    response = requests.get(f"{get_api_endpoint()}get_available_data")
-    assert response.status_code == 200
-    response = requests.get(f"{get_api_endpoint()}get_available_data")
-    assert response.status_code == 200
-
-    time.sleep(3)  # allow remaining data to pass through subprocesses
-
-    # Second run
-    response = requests.get(f"{get_api_endpoint()}start_managed_acquisition")
-    assert response.status_code == 200
-    assert (
-        system_state_eventually_equals(
-            LIVE_VIEW_ACTIVE_STATE, LIVE_VIEW_ACTIVE_WAIT_TIME
-        )
-        is True
-    )
-    response = requests.get(f"{get_api_endpoint()}stop_managed_acquisition")
-    assert response.status_code == 200
-    assert (
-        system_state_eventually_equals(
-            CALIBRATED_STATE, STOP_MANAGED_ACQUISITION_WAIT_TIME
-        )
-        is True
-    )
-
-    test_process_manager.hard_stop_and_join_processes()
 
 
 @pytest.mark.slow
